First iteration pretty print

+++ src/search/tree/tree_base.rs
@@ -34,7 +34,7 @@ impl Index<NodeIndex> for Tree {
 impl Tree {
     pub fn new(size_in_mb: i32) -> Self {
         let bytes = (size_in_mb as usize) * 1024 * 1024;
-        let tree_size = bytes as usize / (48 + 12 * 16);
+        let tree_size = bytes / (48 + 12 * 16);
         let segment_size = (tree_size / SEGMENT_COUNT).min(0x7FFFFFFE);
         let segments = [
             TreeSegment::new(segment_size, 0),
